requirements.txt
Django==1.10.2
django-cors-headers==1.2.1
django-extensions==1.7.4
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
pep8==1.7.0
psycopg2==2.6.2
requests==2.11.1
python-social-auth
algoliasearch
rq
pusher
sendgrid
redis
django-rq
django-cors-middleware